+++ src/frameworks/foundation/ns_file_handle.rs
@@ -104,7 +104,7 @@ pub const CLASSES: ClassExports = objc_classes! {
     match posix_io::read(env, fd, buffer, length) {
         -1 => panic!("readDataOfLength: failed"),
         bytes_read => {
-            assert_eq!(length, bytes_read.try_into().unwrap());
+            assert_eq!(length, NSUInteger::try_from(bytes_read).unwrap());
             msg_class![env; NSData dataWithBytesNoCopy:buffer length:length]
         }
     }